salvar estatistica

+++ backend/Master/Entity/Dto/Response/Domain/Prequal/DtoResponsePrequalSolicitacoes.cs
@@ -8,11 +8,11 @@ namespace Master.Entity.Dto.Response.Domain.Prequal
     [ExcludeFromCodeCoverage]
     public class DtoResponsePrequalSolicitacoes : DtoResponsePrequalSolicitacoesNode
     {
-        public long milis { get; set; }
-        public long totalProcessamentos { get; set; }
+        public int milis { get; set; }
+        public int totalProcessamentos { get; set; }
         public double msPerItem { get; set; }
-        public long totalQualificadas { get; set; }
-        public long totalRejeitadas { get; set; }
+        public int totalQualificadas { get; set; }
+        public int totalRejeitadas { get; set; }
         public double pctPreQualificacao { get; set; }
 
     }
